Move NavLoader to its own file #5557
